requirements.txt
asgiref==3.7.2
click==8.1.7
Django==4.2
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
pep8-naming==0.13.3
platformdirs==3.10.0
pycodestyle==2.11.0
pyflakes==3.1.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
uritemplate==4.1.1
